+++ Tests/NukeTests/ImagePipelineTests/ImagePipelineAsyncAwaitTests.swift
@@ -265,6 +265,7 @@ class ImagePipelineAsyncAwaitTests: XCTestCase, @unchecked Sendable {
         pipeline = pipeline.reconfigured {
             $0.dataLoader = dataLoader
             $0.isProgressiveDecodingEnabled = true
+            $0.progressiveDecodingInterval = 0
         }
 
         // WHEN
@@ -392,6 +393,7 @@ class ImagePipelineAsyncAwaitTests: XCTestCase, @unchecked Sendable {
         pipeline = pipeline.reconfigured {
             $0.dataLoader = dataLoader
             $0.isProgressiveDecodingEnabled = true
+            $0.progressiveDecodingInterval = 0
         }
 
         // WHEN